chore(PocketIC): split tests of IcpFeatures into a separate target
